+++ src/aui-scheduler/js/aui-scheduler-view-day.js
@@ -55,6 +55,7 @@ var Lang = A.Lang,
 
     CSS_SCHEDULER_EVENT = getCN('scheduler-event'),
     CSS_SCHEDULER_EVENT_DISABLED = getCN('scheduler-event', 'disabled'),
+    CSS_SCHEDULER_EVENT_INTERSECTING = getCN('scheduler-event', 'intersecting'),
     CSS_SCHEDULER_EVENT_PROXY = getCN('scheduler-event', 'proxy'),
 
     CSS_SCHEDULER_TODAY = getCN('scheduler', 'today'),
@@ -67,6 +68,8 @@ var Lang = A.Lang,
     CSS_SCHEDULER_VIEW_DAY_HEADER_COL = getCN('scheduler-view', 'day', 'header', 'col'),
     CSS_SCHEDULER_VIEW_DAY_HEADER_DAY = getCN('scheduler-view', 'day', 'header', 'day'),
     CSS_SCHEDULER_VIEW_DAY_HEADER_DAY_FIRST = getCN('scheduler-view', 'day', 'header', 'day', 'first'),
+    CSS_SCHEDULER_VIEW_DAY_HEADER_DAY_NUMBER = getCN('scheduler-view', 'day', 'header', 'day', 'number'),
+    CSS_SCHEDULER_VIEW_DAY_HEADER_DAY_WEEKDAY = getCN('scheduler-view', 'day', 'header', 'day', 'weekday'),
     CSS_SCHEDULER_VIEW_DAY_HEADER_TABLE = getCN('scheduler-view', 'day', 'header', 'table'),
     CSS_SCHEDULER_VIEW_DAY_HEADER_VIEW_LABEL = getCN('scheduler-view', 'day', 'header', 'view', 'label'),
     CSS_SCHEDULER_VIEW_DAY_ICON_GRIP_HORIZONTAL = getCN('scheduler-view', 'icon', 'grip', 'horizontal'),
@@ -130,6 +133,8 @@ var Lang = A.Lang,
 
     TPL_SCHEDULER_VIEW_DAY_HEADER_DAY = '<th class="' + CSS_SCHEDULER_VIEW_DAY_HEADER_DAY +
         '" data-colnumber="{colNumber}"><a href="#">&nbsp;</a></th>',
+    TPL_SCHEDULER_VIEW_DAY_HEADER_DAY_CONTENT = '<span class="' + CSS_SCHEDULER_VIEW_DAY_HEADER_DAY_NUMBER +
+        '">%d</span> <span class="' + CSS_SCHEDULER_VIEW_DAY_HEADER_DAY_WEEKDAY + '">%a</span>',
     TPL_SCHEDULER_VIEW_DAY_HEADER_DAY_FIRST = '<td class="' + [CSS_SCHEDULER_VIEW_DAY_HEADER_DAY,
         CSS_SCHEDULER_VIEW_DAY_HEADER_DAY_FIRST].join(' ') + '"></td>';
 
@@ -250,7 +255,7 @@ var SchedulerDayView = A.Component.create({
 
                 return A.DataType.Date.format(
                     date, {
-                        format: '<span>%d</span> %a',
+                        format: TPL_SCHEDULER_VIEW_DAY_HEADER_DAY_CONTENT,
                         locale: scheduler.get('locale')
                     }
                 );
@@ -544,22 +549,7 @@ var SchedulerDayView = A.Component.create({
             instance.headerTableNode.delegate(
                 'click', A.bind(instance._onClickDaysHeader, instance), '.' + CSS_SCHEDULER_VIEW_DAY_HEADER_DAY);
 
-            instance.columnData.delegate(
-                'mousedown', A.bind(instance._onMouseDownTableCol, instance), '.' +
-                CSS_SCHEDULER_VIEW_DAY_TABLE_COL);
-
-            instance.columnData.delegate(
-                'mouseenter', A.bind(instance._onMouseEnterEvent, instance), '.' + CSS_SCHEDULER_EVENT);
-
-            instance.columnData.delegate(
-                'mouseleave', A.bind(instance._onMouseLeaveEvent, instance), '.' + CSS_SCHEDULER_EVENT);
-
-            instance.columnData.delegate(
-                'mousemove', A.bind(instance._onMouseMoveTableCol, instance), '.' +
-                CSS_SCHEDULER_VIEW_DAY_TABLE_COLDAY);
-
-            instance.columnData.delegate(
-                'mouseup', A.bind(instance._onMouseUpTableCol, instance), '.' + CSS_SCHEDULER_VIEW_DAY_TABLE_COL);
+            this._bindMouseEvents();
 
             instance.on('drag:end', instance._onEventDragEnd);
             instance.on('drag:start', instance._onEventDragStart);
@@ -895,6 +885,13 @@ var SchedulerDayView = A.Component.create({
                     evtNode.setStyle('width', width + '%');
                     evtNode.setStyle('left', left + '%');
 
+                    if (total > 1) {
+                        evtNode.addClass(CSS_SCHEDULER_EVENT_INTERSECTING);
+                    }
+                    else {
+                        evtNode.removeClass(CSS_SCHEDULER_EVENT_INTERSECTING);
+                    }
+
                     var evtParentNode = evtNode.get('parentNode');
 
                     if (evtParentNode) {
@@ -1076,6 +1073,31 @@ var SchedulerDayView = A.Component.create({
         },
 
         /**
+         * Binds the events related to the mouse.
+         *
+         * @method _bindMouseEvents
+         * @protected
+         */
+        _bindMouseEvents: function() {
+            this.columnData.delegate(
+                'mousedown', A.bind(this._onMouseDownTableCol, this), '.' +
+                CSS_SCHEDULER_VIEW_DAY_TABLE_COL);
+
+            this.columnData.delegate(
+                'mouseenter', A.bind(this._onMouseEnterEvent, this), '.' + CSS_SCHEDULER_EVENT);
+
+            this.columnData.delegate(
+                'mouseleave', A.bind(this._onMouseLeaveEvent, this), '.' + CSS_SCHEDULER_EVENT);
+
+            this.columnData.delegate(
+                'mousemove', A.bind(this._onMouseMoveTableCol, this), '.' +
+                CSS_SCHEDULER_VIEW_DAY_TABLE_COLDAY);
+
+            this.columnData.delegate(
+                'mouseup', A.bind(this._onMouseUpTableCol, this), '.' + CSS_SCHEDULER_VIEW_DAY_TABLE_COL);
+        },
+
+        /**
          * Aligns the dragging `SchedulerEvent` to the X axis while bound to the
          * Y axis on the `activeColumn`.
          *
@@ -1143,6 +1165,17 @@ var SchedulerDayView = A.Component.create({
 
                 instance.plotEvent(placeholder);
             }
+        },
+
+        /**
+         * Finds and returns the column related to the given event.
+         *
+         * @method _findActiveColumn
+         * @param {EventFacade} event
+         * @protected
+         */
+        _findActiveColumn: function(event) {
+            return event.currentTarget;
         },
 
         /**
@@ -1317,31 +1350,7 @@ var SchedulerDayView = A.Component.create({
                 recorder.hidePopover();
 
                 if (target.test('.' + CSS_SCHEDULER_VIEW_DAY_TABLE_COL_SHIM)) {
-                    instance.startXY = [event.pageX, event.pageY];
-
-                    var colNumber = toNumber(event.currentTarget.attr('data-colnumber'));
-                    var startDate = instance.getDateByColumn(colNumber);
-                    var clickLeftTop = instance.getXYDelta(event);
-
-                    instance.roundToNearestHour(
-                        startDate, instance.getYCoordTime(clickLeftTop[1]));
-
-                    var endDate = DateMath.add(startDate, DateMath.MINUTES, recorder.get('duration'));
-
-                    recorder.move(startDate, {
-                        silent: true
-                    });
-
-                    recorder.setAttrs({
-                        allDay: false,
-                        endDate: endDate
-                    }, {
-                        silent: true
-                    });
-
-                    instance.creationStartDate = startDate;
-
-                    event.halt();
+                    this._prepareEventCreation(event);
                 }
                 else if (target.test(
                             ['.' + CSS_SCHEDULER_VIEW_DAY_RESIZER,
@@ -1395,7 +1404,7 @@ var SchedulerDayView = A.Component.create({
          */
         _onMouseMoveTableCol: function(event) {
             var instance = this;
-            var activeColumn = event.currentTarget;
+            var activeColumn = this._findActiveColumn(event);
             var recorder = instance.get('scheduler').get('eventRecorder');
 
             if (instance.activeColumn !== activeColumn) {
@@ -1475,6 +1484,44 @@ var SchedulerDayView = A.Component.create({
             if (instance.headerView) {
                 instance.headerView.set('scheduler', event.newVal);
             }
+        },
+
+        /**
+         * Prepares for the creation of a new scheduler event.
+         *
+         * @method _prepareEventCreation
+         * @protected
+         */
+        _prepareEventCreation: function(event, duration) {
+            var clickLeftTop = this.getXYDelta(event),
+                colNumber = toNumber(event.currentTarget.attr('data-colnumber')),
+                endDate,
+                startDate = this.getDateByColumn(colNumber),
+                recorder = this.get('scheduler').get('eventRecorder');
+
+            this.startXY = [event.pageX, event.pageY];
+
+            this.roundToNearestHour(startDate, this.getYCoordTime(clickLeftTop[1]));
+
+            if (!duration) {
+                duration = recorder.get('duration');
+            }
+            endDate = DateMath.add(startDate, DateMath.MINUTES, duration);
+
+            recorder.move(startDate, {
+                silent: true
+            });
+
+            recorder.setAttrs({
+                allDay: false,
+                endDate: endDate
+            }, {
+                silent: true
+            });
+
+            this.creationStartDate = startDate;
+
+            event.halt();
         },
 
         /**