+++ src/aui-scheduler/HISTORY.md
@@ -5,6 +5,7 @@
 ## @VERSION@
 
 * [AUI-1370](https://issues.liferay.com/browse/AUI-1370) Add ability to set an event's border color, style, and width
+* [AUI-1309](https://issues.liferay.com/browse/AUI-1309) Make aui-scheduler responsive
 
 ## [2.5.0](https://github.com/liferay/alloy-ui/releases/tag/2.5.0)